eliminate lat/lon type conversion in osm (define as double)

+++ htroot/osm.java
@@ -15,15 +15,15 @@ public class osm {
     public static RasterPlotter respond(@SuppressWarnings("unused") final RequestHeader header, final serverObjects post, @SuppressWarnings("unused") final serverSwitch env) {
 
         int zoom = 10;
-        float lat = 50.11670f;
-        float lon = 8.68333f;
+        double lat = 50.11670d;
+        double lon = 8.68333d;
         int width = 3;
         int height = 3;
 
         if (post != null) {
             zoom = post.getInt("zoom", zoom);
-            lat = post.getFloat("lat", lat);
-            lon = post.getFloat("lon", lon);
+            lat = post.getDouble("lat", lat);
+            lon = post.getDouble("lon", lon);
             width = post.getInt("width", width);
             height = post.getInt("height", height);
         }
